+++ src/models/command.ts
@@ -1,4 +1,4 @@
-import { Client, Message } from "discord.js";
+import { Client, Message, Permissions } from "discord.js";
 
 /** The template for all commands to ensure required properties */
 export default interface Command {
@@ -6,6 +6,9 @@ export default interface Command {
     readonly description: string;
     readonly syntax: string;
     readonly aliases?: string[];
+
+    readonly perms: Permissions[];
+    readonly guildOnly: Boolean;
     
     execute(message: Message, args: string[], client: Client): void;
 }