+++ src/util/array_helper.ts
@@ -1,0 +1,5 @@
+import { PermissionString, Permissions } from "discord.js";
+
+export function PermissionsFrom(...arr: PermissionString[]): Permissions[] {
+    return arr.map(perm => new Permissions(perm));
+}